+++ perllib/FixMyStreet/DB/Result/Comment.pm
@@ -229,13 +229,17 @@ sub meta_line {
 
     if ($self->anonymous or !$self->name) {
         $meta = sprintf( _( 'Posted anonymously at %s' ), Utils::prettify_dt( $self->confirmed ) )
-    } elsif ($self->user->from_body) {
+    } elsif ($self->user->from_body || ( $self->name eq _('admin') && $self->user->is_superuser)) {
         my $user_name = FixMyStreet::Template::html_filter($self->user->name);
         my $body = $self->user->body;
         if ($body eq 'Bromley Council') {
             $body = "$body <img src='/cobrands/bromley/favicon.png' alt=''>";
         } elsif ($body eq 'Royal Borough of Greenwich') {
             $body = "$body <img src='/cobrands/greenwich/favicon.png' alt=''>";
+        }
+        if ($self->name eq _('admin') && !$self->user->from_body) {
+            $c->log->debug('admin');
+            $body = _('admin');
         }
         my $can_view_contribute = $c->user_exists && $c->user->has_permission_to('view_body_contribute_details', $self->problem->bodies_str_ids);
         if ($self->text) {